feat: Implement ip cleaning feature ✨
Clean the IPs from a changelog older than $conf['recent_days'].
They are replaced by whitespace to ensure the same byte length when
writing it back to the file.
Since changelogs could be very(!) large, they are read line-by-line and
we store the position where we left of, so it can be used for the next
run.

ADI-169